Emit StorageLive/StorageDead statements for MIR temporaries
When we compute which locals are live across a yield point,
we rely on StorageLive/StorageDead statements to help determine
if a variable is live or not.

Previously, we did not emit these
statements for the temporary local create for a statement
(e.g. 'String::new()'). This lead to these temporaries being incorrectly
determined to be live across a yield point, when the earlier AST
analysis had correctly concluded that they could not possibly live
across any yield point.